tdl_sdk: copy lib

+++ tdl_sdk/build_tdl_sdk.sh
@@ -142,7 +142,9 @@ popd
 # build end
 
 # copy lib
-mkdir -p "$SYSTEM_OUT_DIR"/lib/
-mkdir -p "$SYSTEM_OUT_DIR"/usr/bin/ai
-cp -arf "$TDL_SDK_INSTALL_PATH"/lib/*.so* "$SYSTEM_OUT_DIR"/lib/
-cp -arf "$TDL_SDK_INSTALL_PATH"/bin/* "$SYSTEM_OUT_DIR"/usr/bin/ai/
+if [ -f "${SYSTEM_OUT_DIR}" ]; then
+  mkdir -p "$SYSTEM_OUT_DIR"/lib/
+  mkdir -p "$SYSTEM_OUT_DIR"/usr/bin/ai
+  cp -arf "$TDL_SDK_INSTALL_PATH"/lib/*.so* "$SYSTEM_OUT_DIR"/lib/
+  cp -arf "$TDL_SDK_INSTALL_PATH"/bin/* "$SYSTEM_OUT_DIR"/usr/bin/ai/
+fi
